336 verified.
git-svn-id: http://tradelink.googlecode.com/svn/trunk@1430 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ TradeLinkAPI/Bar.cs
@@ -16,7 +16,7 @@ namespace TradeLink.API
         decimal Low { get;  }
         decimal Open { get;  }
         decimal Close { get;  }
-        int Volume { get;  }
+        long Volume { get;  }
         bool isNew { get; set;}
         int Bartime { get;  }
         int Bardate { get; }
